add prompt for overwrite file with similar name

+++ distant_operator/src/lib.rs
@@ -1,13 +1,16 @@
-use std::{fs, path::PathBuf, sync::LazyLock};
+use std::{fs::{self, DirEntry}, path::PathBuf, sync::LazyLock};
 
+use anyhow::bail;
 use logging::Logger;
+use nu_ansi_term::Style;
+use promkit::preset::{confirm::Confirm, listbox::Listbox};
 use reedline::ExternalPrinter;
 use serde::{Deserialize, Serialize};
 
 pub mod logging;
 
 // internal
-static LOGGER: Logger = Logger;
+pub static LOGGER: Logger = Logger;
 pub static PRINTER: LazyLock<ExternalPrinter<String>> = LazyLock::new(|| ExternalPrinter::new(8));
 
 // directories
@@ -35,8 +38,8 @@ pub struct Profile {
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct ProfileInner {
-    address: String,
-    port: u16,
+    pub address: String,
+    pub port: u16,
     public_key: String,
     private_key: String,
 }
@@ -55,15 +58,56 @@ impl Profile {
     }
 
     pub fn write_to(&self, path: PathBuf) -> Result<(), anyhow::Error> {
+        if fs::exists(path.clone()).unwrap_or_default() {
+            if !Confirm::new(format!(
+                "The file {:?} already exists; Do you want to replace it?",
+                path.file_name().unwrap()
+            ))
+            .prompt()?
+            .run()?
+            .to_lowercase()
+            .contains('y')
+            {
+                return Ok(());
+            };
+        }
+
         let str = toml::to_string(&self.inner)?;
         fs::write(path, str)?;
         Ok(())
     }
 }
 
-pub fn init() -> Result<(), anyhow::Error> {
-    log::set_logger(&LOGGER)?;
-    log::set_max_level(log::LevelFilter::Debug);
+pub fn select_profile() -> Result<Profile, anyhow::Error> {
+    let path = PROFILE_DIR.clone();
 
-    Ok(())
+    let profiles: Vec<String> = fs::read_dir(path)?
+        .filter(|v| v.is_ok())
+        .map(|v| format_profile_name(v.unwrap()))
+        .collect();
+
+    if profiles.len() == 0 {
+        bail!(
+            "No profiles found in {:?}\n {}: run the `import` command to add a profile",
+            PROFILE_DIR.clone(),
+            Style::new().bold().paint("Note")
+        )
+    }
+
+    let selected = Listbox::new(profiles)
+        .title("Select A Profile")
+        .listbox_lines(5)
+        .prompt()?
+        .run()?;
+    let file = selected.split(" | ").next().unwrap();
+
+    Profile::parse(PROFILE_DIR.join(file))
+}
+
+fn format_profile_name(file: DirEntry) -> String {
+    let file_name = file.file_name().into_string().unwrap();
+    let str = fs::read_to_string(file.path()).unwrap();
+    let parsed: ProfileInner = toml::from_str(&str).unwrap();
+
+    format!("{} | {}", file_name, parsed.address)
 }
